# openapi.yaml
openapi: 3.1.0
info:
  title: CP-Tekniker Device Management API
  version: 1.0.0
  description: |
    API-dokumentasjon for CP-Tekniker Device Management App. Dette dekker alle publiserte (backend) endepunkter og de viktigste eksterne ClearPass-API-endepunktene som brukes av backend.
servers:
  - url: http://localhost:8000
    description: Lokal utviklingsserver
paths:
...
  /get_device_info:
    get:
      summary: Hent enhetsinformasjon
      description: Hent informasjon om en enhet basert på MAC-adresse (krever innlogging).
      parameters:
        - in: query
          name: macaddr
          schema:
            type: string
          required: true
          description: MAC-adresse til enheten
      responses:
        '200':
          description: Enhetsinformasjon
          content:
            application/json:
              schema:
                type: object
        '400':
          description: MAC-adresse er påkrevd
        '401':
          description: Ikke autentisert
        '500':
          description: Feil ved henting av enhetsinformasjon
  /create_device:
    post:
      summary: Opprett enhet
      description: Opprett en ny enhet (krever innlogging).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mac:
                  type: string
                  description: MAC-adresse
                role_id:
                  type: string
                  description: Rolle-ID
                enabled:
                  type: boolean
                visitor_name:
                  type: string
                expire_time:
                  type: integer
                  description: Unix-tid (sekunder)
                sponsor_name:
                  type: string
                sponsor_profile:
                  type: string
      responses:
        '201':
          description: Enhet opprettet
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Manglende påkrevde felt
        '401':
          description: Ikke autentisert
        '500':
          description: Feil ved opprettelse av enhet
  /update_device:
    patch:
      summary: Oppdater enhet
      description: Oppdater eksisterende enhet (krever innlogging).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mac:
                  type: string
                  description: MAC-adresse (påkrevd for identifikasjon)
                role_id:
                  type: string
                  description: Rolle-ID
                enabled:
                  type: boolean
                visitor_name:
                  type: string
                expire_time:
                  type: integer
                  description: Unix-tid (sekunder)
                sponsor_name:
                  type: string
                sponsor_profile:
                  type: string
      responses:
        '200':
          description: Enhet oppdatert
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Manglende eller ugyldige felt
        '401':
          description: Ikke autentisert
        '404':
          description: Enhet ikke funnet
        '500':
          description: Feil ved oppdatering av enhet
...